fix common transform unit tests
