+++ mmpose/datasets/transforms/__init__.py
@@ -1,16 +1,18 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .bottomup_transforms import (BottomupGetHeatmapMask, BottomupRandomAffine,
                                   BottomupResize)
-from .common_transforms import (Albumentation, GetBBoxCenterScale,
-                                PhotometricDistortion, RandomBBoxTransform,
-                                RandomFlip, RandomHalfBody)
+from .common_transforms import (Albumentation, GenerateTarget,
+                                GetBBoxCenterScale, PhotometricDistortion,
+                                RandomBBoxTransform, RandomFlip,
+                                RandomHalfBody)
 from .formatting import PackPoseInputs
 from .loading import LoadImage
-from .topdown_transforms import TopdownAffine, TopdownGenerateTarget
+from .topdown_transforms import TopdownAffine
 
 __all__ = [
     'GetBBoxCenterScale', 'RandomBBoxTransform', 'RandomFlip',
-    'RandomHalfBody', 'TopdownAffine', 'TopdownGenerateTarget',
-    'Albumentation', 'PhotometricDistortion', 'PackPoseInputs', 'LoadImage',
-    'BottomupGetHeatmapMask', 'BottomupRandomAffine', 'BottomupResize'
+    'RandomHalfBody', 'TopdownAffine', 'Albumentation',
+    'PhotometricDistortion', 'PackPoseInputs', 'LoadImage',
+    'BottomupGetHeatmapMask', 'BottomupRandomAffine', 'BottomupResize',
+    'GenerateTarget'
 ]
